+++ types/api_UserInfo.go
@@ -5,24 +5,24 @@ type UserInfoData struct {
 	RawJson   string
 	Level     int `json:"level"`
 	UserPoint struct {
-		Balance      int   `json:"balance"`
-		BlockBalance int   `json:"blockBalance"`
-		Status       int   `json:"status"`
-		UpdateTime   int64 `json:"updateTime"`
-		UserId       int   `json:"userId"`
-		Version      int   `json:"version"`
+		Balance      int `json:"balance"`
+		BlockBalance int `json:"blockBalance"`
+		Status       int `json:"status"`
+		UpdateTime   int `json:"updateTime"`
+		UserId       int `json:"userId"`
+		Version      int `json:"version"`
 	} `json:"userPoint"`
 	MobileSign       bool   `json:"mobileSign"`
 	PcSign           bool   `json:"pcSign"`
 	Viptype          int    `json:"viptype"`
-	Expiretime       int64  `json:"expiretime"`
-	BackupExpireTime int64  `json:"backupExpireTime"`
+	Expiretime       int    `json:"expiretime"`
+	BackupExpireTime int    `json:"backupExpireTime"`
 	Storeurl         string `json:"storeurl"`
 	MallDesc         string `json:"mallDesc"`
 	StoreTitle       string `json:"storeTitle"`
 	Pubwords         string `json:"pubwords"`
 	GameConfig       struct {
-		ExpirationTime       int64  `json:"expirationTime"`
+		ExpirationTime       int    `json:"expirationTime"`
 		GameCenterEntryTitle string `json:"gameCenterEntryTitle"`
 		GameCenterPic        string `json:"gameCenterPic"`
 		GameCenterReddot     int    `json:"gameCenterReddot"`
